# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LMS Backend Core API",
        "description": "Learning Management System Backend API Documentation. This API provides comprehensive endpoints for managing courses, users, enrollments, and learning progress.",
        "contact": {
            "name": "LMS API Support",
            "email": "api-support@lms.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "LMS API Server"
        }
    ],
    "paths": {
...
        "/api/profile": {
            "get": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Get user profile",
                "description": "Get the authenticated user's profile information",
                "operationId": "85d739158552477973430d9c6aaf191e",
                "responses": {
                    "200": {
                        "description": "Profile retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Profile retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "john@example.com"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "+1234567890"
                                                },
                                                "bio": {
                                                    "type": "string",
                                                    "example": "Software developer with 5 years experience"
                                                },
                                                "avatar_url": {
                                                    "type": "string",
                                                    "example": "https://example.com/avatars/user1.jpg"
                                                },
                                                "date_of_birth": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "1990-01-15"
                                                },
                                                "gender": {
                                                    "type": "string",
                                                    "example": "male"
                                                },
                                                "country": {
                                                    "type": "string",
                                                    "example": "United States"
                                                },
                                                "timezone": {
                                                    "type": "string",
                                                    "example": "America/New_York"
                                                },
                                                "language": {
                                                    "type": "string",
                                                    "example": "en"
                                                },
                                                "email_verified_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "stats": {
                                                    "properties": {
                                                        "total_enrollments": {
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "completed_courses": {
                                                            "type": "integer",
                                                            "example": 3
                                                        },
                                                        "active_enrollments": {
                                                            "type": "integer",
                                                            "example": 2
                                                        },
                                                        "certificates_earned": {
                                                            "type": "integer",
                                                            "example": 3
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Update user profile",
                "description": "Update the authenticated user's profile information",
                "operationId": "d6ec7eead79411b2060e10d562bd7a87",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Full name",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "phone": {
                                        "description": "Phone number",
                                        "type": "string",
                                        "example": "+1234567890"
                                    },
                                    "bio": {
                                        "description": "Biography",
                                        "type": "string",
                                        "example": "Software developer with 5 years experience"
                                    },
                                    "date_of_birth": {
                                        "description": "Date of birth",
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-15"
                                    },
                                    "gender": {
                                        "description": "Gender",
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female",
                                            "other",
                                            "prefer_not_to_say"
                                        ],
                                        "example": "male"
                                    },
                                    "country": {
                                        "description": "Country",
                                        "type": "string",
                                        "example": "United States"
                                    },
                                    "timezone": {
                                        "description": "Timezone",
                                        "type": "string",
                                        "example": "America/New_York"
                                    },
                                    "language": {
                                        "description": "Preferred language",
                                        "type": "string",
                                        "example": "en"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Profile updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profile/password": {
            "put": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Change password",
                "description": "Change the authenticated user's password",
                "operationId": "3747143c711f492aef2a03ff9ce3241b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "current_password": {
                                        "description": "Current password",
                                        "type": "string",
                                        "example": "oldpassword123"
                                    },
                                    "password": {
                                        "description": "New password",
                                        "type": "string",
                                        "example": "newpassword123"
                                    },
                                    "password_confirmation": {
                                        "description": "Confirm new password",
                                        "type": "string",
                                        "example": "newpassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password changed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Password changed successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error or incorrect current password"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "bio": {
                        "type": "string",
                        "example": "Software developer with 5 years experience"
                    },
                    "avatar_url": {
                        "type": "string",
                        "example": "https://example.com/avatars/user1.jpg"
                    },
                    "date_of_birth": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-01-15"
                    },
                    "gender": {
                        "type": "string",
                        "example": "male"
                    },
                    "country": {
                        "type": "string",
                        "example": "United States"
                    },
                    "timezone": {
                        "type": "string",
                        "example": "America/New_York"
                    },
                    "language": {
                        "type": "string",
                        "example": "en"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter JWT Bearer token in format: Bearer <token>"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}